Move `try_connection` helper to `client` module
